# [prototype runtime] Optimize performance